+++ include/vyb/semantic.hpp
@@ -406,6 +406,7 @@ public:
     void visit(ast::ArrayInitializationExpression* node) override;
     void visit(ast::TypeofExpression* node) override;
     void visit(ast::TypenameExpression* node) override;
+    void visit(ast::AsExpression* node) override;
 
     // Declarations
     void visit(ast::Module* node) override; // Added declaration